+++ Bootstrap/Managers/InternalCalls.cpp
@@ -42,6 +42,7 @@ void InternalCalls::MelonLogger::AddInternalCalls()
 
 #pragma region MelonUtils
 bool InternalCalls::MelonUtils::IsGameIl2Cpp() { return Game::IsIl2Cpp; }
+bool InternalCalls::MelonUtils::IsOldMono() { return Mono::IsOldMono; }
 Mono::String* InternalCalls::MelonUtils::GetApplicationPath() { return Mono::Exports::mono_string_new(Mono::domain, Game::ApplicationPath); }
 Mono::String* InternalCalls::MelonUtils::GetGameName() { return Mono::Exports::mono_string_new(Mono::domain, Game::Name); }
 Mono::String* InternalCalls::MelonUtils::GetGameDeveloper() { return Mono::Exports::mono_string_new(Mono::domain, Game::Developer); }
@@ -53,6 +54,7 @@ void InternalCalls::MelonUtils::SCT(Mono::String* title) { if (title == NULL) re
 void InternalCalls::MelonUtils::AddInternalCalls()
 {
 	Mono::AddInternalCall("MelonLoader.MelonUtils::IsGameIl2Cpp", IsGameIl2Cpp);
+	Mono::AddInternalCall("MelonLoader.MelonUtils::IsOldMono", IsOldMono);
 	Mono::AddInternalCall("MelonLoader.MelonUtils::GetApplicationPath", GetApplicationPath);
 	Mono::AddInternalCall("MelonLoader.MelonUtils::GetGameName", GetGameName);
 	Mono::AddInternalCall("MelonLoader.MelonUtils::GetGameDeveloper", GetGameDeveloper);