+++ MelonLoader/Melons/Utils.cs
@@ -22,6 +22,8 @@ namespace MelonLoader
         [MethodImpl(MethodImplOptions.InternalCall)]
         public extern static bool IsGameIl2Cpp();
         [MethodImpl(MethodImplOptions.InternalCall)]
+        public extern static bool IsOldMono();
+        [MethodImpl(MethodImplOptions.InternalCall)]
         [return: MarshalAs(UnmanagedType.LPStr)]
         public extern static string GetApplicationPath();
         [MethodImpl(MethodImplOptions.InternalCall)]